[BUGFIX] Avoid template exception for wrong sessions
If the session does not return the current step from
a session value, then formhandler should use »1« by default
and make some calculations afterwards.

The "view subpart" method however ignores the fallback and
always tries to get the value from the session. This will
lead to an exception when a form is ubmitted on another domain
(submitted=1), all values are valid (valid=1) but the form
is still considered nit finished (finished=false), due to
the wrong session value.

Use the calculated step instead.